[Fix] Change redirect when user is not authorized, add toast
