+++ helpers/index.js
@@ -2,3 +2,4 @@ export { getErrorMessage } from './getErrorMessage';
 export { formatDate, fromNow } from './date';
 export { shortify } from './shortify';
 export { uploadImage } from './uploadImage';
+export { disposeQueryMessage } from './disposeQueryMessage';
